service/requirements.txt
fastapi==0.115.12
numpy==2.2.6
opencv_python==4.11.0.86
Pillow==11.2.1
torch==2.7.0+cpu
transformers==4.51.3
